requirements.txt
Django==2.2b1
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
djoser==1.5.0
